No changes to Makefile. Implemented functions across the board (ba dum tss) for controlling the timer on the game screen.
To accomplish this, I am using Glib::signal_timeout().connect(...) to call a timeout handling function every second. To match the required form of the timeout handler, I have the handler in main.cpp calling a helper function in the board class. The helper function returns a formatted string as XX Minutes YY seconds that can be directly placed in the current_time_time_label .

I also created a private method in the board class that takes a long representing a number of seconds elapsed and returns a formatted string in XX Minutes YY Seconds. I created this function because the same code would have showed up twice, and that's what functions are for.

The current time functionality still has some bugs. Notably, if the game is closed and re-opened, the current time jumps way up. This will be addressed.