Added two new private variables to the board class. The first is an input/output stream variable to hold the user data file. The filename will be created as username_data.txt, and therefore this variable will be set based on the username that the user enters before beginning. A second variable was added to the board class to hold the username.
The GUI was altered to add an initial, pre-main menu screen to collect the user's name.

A header was added to and removed from board.cpp.

The new window and the new button were added to the main function for stylesheets and eventually funcitonality. The entry for the user's name has not yet been added.