No real change in GUI.glade. Just messing around.
Added flag to include glibmm/ustring.h in the header file for the board class.

Changed getter and setter for username in board class to use Glibmm::ustring instead of std::string since they are essentially the same, except that the former is cross-compatible.

Implemented functionality for entering the username. The player info dialog spawns before the game will open and requests a name. If a name is not entered, the user cannot submit. If the users exits the dialog and tries again, the dialog will reopen until a username is entered.

After entering a username, the dialog will not appear again.