Assigned an ID to the new game button in the congratulations dialog.
Added two new methods and one new variable to the board class. One method resets the board internal variables 'blocks' and 'game_board', and the other calls this method and the reset_time method for a convenient way to do a full reset of the game state. The new variable is simply two-dimensional vector of zeros to easily reset the 'blocks' and 'game_board' variables.

Renamed new_game function to new_game_from_paused to separate it from a different new_game function. Connected signals and CSS to the new game button in the congratulations dialog.