Added and removed print statements for board.cpp .
Changed close_game function to take a boolean as an argument representing whether or not a game is in progress. If it is, close_game will set the total time in the board object. If it's not, then the button must have been clicked from the congratulations dialog, so it hides that dialog instead.

Uncommented signal_clicked connection for the begin button that was commented out for experimentation.