Added functionality to save and load the state of the board using the key file.
Two functions exist in the board class for this. save_board_state saves each row of the board to it's own key in the keyfile. load_board_state does the opposite, and loads each row-key from the keyfile into the game_board internal variable.

After loading data from the keyfile, each cell is set in the initialize_board function to match the internal board representation.