Added a new member variable to the board class to indicate whether or not the reserved cells have had numbers set.
Added a new method call set_reserved to populate the reserved member variable with the proper values from the keyfile if the user is already in a game.

Added a new method generate_reserved to generate the coordinates of reserved cells.

Added a method to return the vector containing the coordinates of reserved cells.

Added a macro to define how many cells should be reserved.