Added method to board class to save the total_time and fastest_time member variables.
Removed calls to generate_reserved across files. There is now a single call to generate_reserved in open_game that is called when a game is opened. Either the reserved_set member variable will or won't be set by this point, and so it is more elegant and more efficient for a single function call to operate here.

This arrangement fixes the issue where reserved cells were being improperly regenerated. Additionally, the new game button works as expected.

A function was added to main.cpp to act as a callback that's called when text is removed from one of the cells. It was found that text being removed was not being handled. This callback is connected in initialize_board, right after the callback for inserted text is connected.