Removed inset border from cells on gameboard. This was done by creating new css classes that specifically set the borders for each cell. In between the cells, there are now black borders.
Reserved cells are now not only uneditable, but also non-sensitive so that they can't be clicked. Background color for reserved cells was set to white.

The main Gtk::Window was switched to a Gtk::ApplicationWindow, which functions largely the same except that the Gtk::ApplicationWindow inherits from Gio::ActionGroup, which will be useful when undo/redo functionality is connected.